+++ dev/openapi-ts.config.ts
@@ -6,9 +6,9 @@ import { customClientPlugin } from '@hey-api/custom-client/plugin';
 // @ts-ignore
 import { defineConfig, utils } from '@hey-api/openapi-ts';
 
-import { getSpecsPath } from '../../utils';
 // @ts-ignore
-import { myClientPlugin } from './custom/client/plugin';
+import { myClientPlugin } from '../packages/openapi-ts-tests/main/test/custom/client/plugin';
+import { getSpecsPath } from '../packages/openapi-ts-tests/utils';
 
 // @ts-ignore
 // eslint-disable-next-line arrow-body-style
@@ -96,14 +96,14 @@ export default defineConfig(() => {
         //   importFileExtension: '.ts',
         //   // indexFile: false,
         //   // lint: 'eslint',
-        //   path: path.resolve(__dirname, 'generated', 'sample'),
+        //   path: path.resolve(__dirname, '.gen'),
         //   tsConfigPath: path.resolve(
         //     __dirname,
         //     'tsconfig',
         //     'tsconfig.nodenext.json',
         //   ),
         // },
-        path.resolve(__dirname, 'generated', 'sample'),
+        '.gen',
       ],
       parser: {
         filters: {
@@ -138,8 +138,8 @@ export default defineConfig(() => {
               // if (!symbol.external && !symbol.meta?.resourceType) {
               //   console.log(`[${plugin.name}]:`, symbol.name);
               // }
-              if (!symbol.external && !symbol.meta?.path) {
-                console.log(`[${plugin.name}]:`, symbol.name, symbol.meta);
+              if (plugin && !symbol.external && !symbol.meta?.path) {
+                // console.log(`[${plugin.name}]:`, symbol.name, symbol.meta);
               }
               // if (symbol.meta?.tags && symbol.meta?.tags.size > 0) {
               //   console.log(
@@ -213,11 +213,11 @@ export default defineConfig(() => {
         // myClientPlugin(),
         {
           // baseUrl: false,
-          exportFromIndex: true,
-          // name: '@hey-api/client-fetch',
+          // exportFromIndex: true,
+          name: '@hey-api/client-fetch',
           // name: 'legacy/angular',
           // runtimeConfigPath: path.resolve(__dirname, 'hey-api.ts'),
-          runtimeConfigPath: './src/hey-api.ts',
+          // runtimeConfigPath: './src/hey-api.ts',
           // strictBaseUrl: true,
           // throwOnError: true,
         },
@@ -269,10 +269,10 @@ export default defineConfig(() => {
           // signature: 'object',
           // transformer: '@hey-api/transformers',
           // transformer: true,
-          validator: {
-            request: 'zod',
-            response: 'zod',
-          },
+          // validator: {
+          //   request: 'zod',
+          //   response: 'zod',
+          // },
           '~hooks': {
             symbols: {
               // getFilePath: (symbol) => {
@@ -396,7 +396,7 @@ export default defineConfig(() => {
           },
           exportFromIndex: true,
           // metadata: true,
-          name: 'zod',
+          // name: 'zod',
           // requests: {
           //   // case: 'SCREAMING_SNAKE_CASE',
           //   // name: 'z{{name}}TestData',
@@ -478,7 +478,7 @@ export default defineConfig(() => {
     //     // level: 'debug',
     //     path: './logs',
     //   },
-    //   output: path.resolve(__dirname, 'generated', 'sample'),
+    //   output: '.gen',
     // },
   ];
 });